save store in localStorage
